Now uses setting in Config to determine how to scrape sponsors.
Philadelphia, Chicago, and other cities have different ways in which they
format sponsors for things in city council. Now you can specify whether
or not the sponsors are listed as links or a list of sponsors separated
by commas.

To scrape legistar instances where sponsors are a list of links
set sponsor_links = True in the Config for the scraper.

If the list is a list separated by commas, set sponsor_links = False.